tests: unit: simplify unit tests layout
Move modules unit tests directly under unit/ and remove unit/harness to
use tests/harness instead.
