+++ tests/unit/core/flavor.c
