+++ tests/unit/core/hook.c
