+++ tests/unit/core/list.c
