+++ tests/unit/core/opts.c
