+++ tests/unit/core/rule.c
