+++ tests/unit/core/verdict.c
